requirements.txt
Django==3.2.4
django_bootstrap_breadcrumbs==0.9.2
django-crispy-forms==1.12.0
django-tables2==2.4.0
Jinja2==3.0.1
Authlib==0.15.4
requests==2.25.1
python-slugify[unidecode]==5.0.2
tablib==3.0.0
hachoir==3.1.2
